src/kit_campus_puzzle/requirements.txt
Django==2.0.3
django-allauth==0.35.0
Pillow==5.0.0
numpy==1.14.2
mysqlclient==1.3.12